[IMPROVE] improve update DOne Task

+++ src/main/java/com/askrindo/service/TaskService.java
@@ -25,5 +25,5 @@ public interface TaskService {
 
     public List<Task> getTaskByReleaseId(String id, String userId, String statusDone);
     public List<Task> getTaskAfterDeadline();
-    public void updateDoneTask(String idTask, Integer prosentase);
+    public void updateDoneTask(String idTask, Float prosentase);
 }
